refactor: 주문 테이블과 관련된 도메인을 별도의 패키지로 분리
