+++ src/main/java/kitchenpos/application/OrderService.java
@@ -8,12 +8,12 @@ import kitchenpos.application.exception.OrderNotFoundException;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.domain.ordertable.repository.OrderTableRepository;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.ui.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.ui.dto.request.CreateOrderRequest;
 import kitchenpos.ui.dto.request.UpdateOrderStatusRequest;
