+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.application.exception.TableGroupNotFoundException;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.TableGroup;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.domain.ordertable.repository.OrderTableRepository;
 import kitchenpos.repository.TableGroupRepository;
 import kitchenpos.ui.dto.request.CreateOrderGroupOrderTableRequest;
 import kitchenpos.ui.dto.request.CreateTableGroupRequest;
