+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -8,6 +8,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import kitchenpos.domain.common.CreatedTimeEntity;
+import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.ordertable.OrderTables;
 
 @Entity
 public class TableGroup extends CreatedTimeEntity {
