+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -17,14 +17,14 @@ import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidOrderLineItemException;
 import kitchenpos.domain.exception.InvalidOrderStatusException;
 import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.domain.ordertable.repository.OrderTableRepository;
 import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.ui.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.ui.dto.request.CreateOrderRequest;
