+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -16,7 +16,7 @@ import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.exception.InvalidOrderStatusCompletionException;
@@ -24,7 +24,7 @@ import kitchenpos.domain.exception.InvalidOrderTableSizeException;
 import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.domain.ordertable.repository.OrderTableRepository;
 import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.repository.TableGroupRepository;
 import kitchenpos.ui.dto.request.CreateOrderGroupOrderTableRequest;
